Move style files to own directory
