+++ src/components/ProcessGraphCanvas.tsx
@@ -24,7 +24,7 @@ import NodeDetails from './NodeDetails';
 import { Edge } from '../types/Edge';
 import { Node } from '../types/Node';
 import InfoBox from './InfoBox';
-import Button from '../styledComponents';
+import Button from '../styles/components';
 
 const Container = styled.div`
   background-color: ${props => props.theme.palette.background.main};
