+++ src/styles/components.ts
